backend/requirements.txt
alembic
uvicorn
fastapi
sqlalchemy
python-dotenv
psycopg2